Use .decode() instead of str() in run_command
Whoops, just converting a byte string to a string leaves you with something
like " b'string\n' " instead of just "string". Need to decode the byte string
to unicode instead of just casting.

Note that for python 3 we could use `str(byteString, 'utf-8'), but that's not
supported by python 2, so use decode instead:
https://docs.python.org/3/library/stdtypes.html#str