Fix bugs and python 3 incompatibilities in chpl_arch
translate(None, punctuation) was used to removed punctuation from a string, but
that's not supported in python3. Instead create a helper remove_punctuation
routine that works for python 2 and 3. There's probably better/faster ways to
remove punctuation, but this was really simple.

I took it from:
http://stackoverflow.com/questions/265960/best-way-to-strip-punctuation-from-a-string-in-python
which has comments about py3 translate not supporting the deletechars argument.

While I was testing this I also found that setting CHPL_HOST_ARCH to "native" or
"unknown" would produce warnings about unsupported archs, so I fixed that.

To repro the unsupported arch warning run:
  `CHPL_HOST_ARCH=native python2 util/chplenv/chpl_arch.py --host`

To repro the python3 translate error run
  `CHPL_HOST_ARCH=native python3 util/chplenv/chpl_arch.py --host`