Add ability to limit how many tests will execute concurrently
Paratest has the ability to run oversubscribed with the `-nodepara` option.
This essentially runs multiple start_tests per machine, which means multiple
compiles/executions can happen simultaneously. However, some chapel programs
(particularly when using qthreads) can use up a significant portion of the
machine, which can lead to timeouts when multiple heavy-weight tests are
running concurrently.

This adds the ability to limit our testing infrastructure so that only one
executable can run at a time. Compilation is single threaded, so this doesn't
limit how many compiles occur and it still lets compilation occur while other
tests are running.

In order to enable this just set `CHPL_TEST_LIMIT_RUNNING_EXECUTABLES=yes`

In order to limit how many executables run at once, we basically use a file as
a mutex. This is race-y, which means occasionally two or more executables might
be running at the same time. It's pretty unlikely this will happen though,
since there's small window for that to happen. And note that the code is
careful to consider cases where paratest is killed and doesn't remove the lock,
or where multiple executables end up running concurrently.

The simplified version of what this code does is:

```python
  while os.path.exists('/tmp/<username>-test-running') and modified < timeout:
      time.sleep(.1)

  # race-y since other process could have created file since while loop
  open('/tmp/<username>-test-running', 'w')

  run_executable()

  os.remove('/tmp/<username>-test-running')
```

Note that we could use fcntl.flock() to prevent the above race, but then if
paratest is killed and we don't get a chance to unlock, we would have to
manually remove the lock, which would be a pain. Python 3.2 added an 'x' mode
to open that supports exclusive creation, but we can't rely on that yet.