Avoid creating a tuple for varargs in some cases
WE USED TO

combine varargs into a tuple the first thing in a function.
E.g when this function:

  proc test(args...) {
    BODY;
  }

is invoked with two formals, the compiler instantiates it into:

  proc test(_e0_args, _e1_args) {
    var args = _build_tuple(2, _e0_args, _e1_args);
    BODY;
  }

Indeed the vararg tuple is created in the callee, not the caller.

WITH THIS COMMIT,

the compiler no longer inserts the _build_tuple() call or creates
the local 'args' tuple when:

(A) BODY does not need 'args' as a whole tuple (see below), and
(B) the "args..." formal's intent is not 'out' or 'inout'.

This results in simpler generated code. Also, as the primary motivation
for this change, if the args formal has a param or type intent, then
(...args) consists of params or types, resp. This latter was not the case
prior to this change.

(A) details: the compiler considers (A) to hold when the only uses
of 'args' in the body, if any, are these:

  (A1) (...args)           --> replaced with _e0_args, _e1_args
  (A2) args(i) for param i --> replaced with _e0_args or _e1_args
  (A3) args.size           --> replaced with 2

The replacements shown are for the test() example above.

(B) details: supporting out/inout intent is left as future work.
These intents require formal temps. So instead of the 'args' tuple
we'd have to create individual temps. Maybe this wouldn't be
much savings to begin with?

IMPLEMENTATION NOTES

When the function whose varargs are being instantiated
had FLAG_PARTIAL_COPY, the replacements needed to be done
separately. The necessary information is passed through
the fields I added on FnSymbol.

While there, I added a missing SET_LINENO and in a different
spot set a node's astloc manually, which is more lightweight
than SET_LINENO.

MULTIPLE VARARGS

are not supported by my changes. This is because FnSymbol
has only a single pair of varargOldFormal and varargNewFormals
fields. One way to support multiple sets of varargs is to
convert the two fields to a map old-formals to their respective
new-formals vectors. Another way is to fall back on the current
create-a-tuple implementation.

Note that the compiler already does not support multiple sets of
varargs. This is mainly because the number of vararg actuals is
computed like so:

  int numCopies = numActuals - workingFn->numFormals() + 1;

If we were, for example, to distinguish different sets of varargs
by their types, we would need to run the entire instantiation
machinery before computing numCopies and instantiating again.

ALSO ADDED TESTS

that exercise varargs in various ways.

Varargs now work with --minimal-modules, too, as long
as (A) and (B) above are satisfied - see minModVarArgs.chpl