Adding module name to the end of generated symbols.
Currently,in the case when two symbols have the same name, they are
distinguished by appending an integer to the end of the symbol name. However
that does make it confusing while reading the code. The changes made here also
append the module name to the end of the generated symbols which makes the
code more readable.

Some points worth making a note:
* In case the symbols are still not unique (even after appending module name,
then we append a unique integer to the end of the symbol for disambiguation.
* Currently for field names, simply appending the module name to the end of
the field name causes some errors, which were not getting easily fixed.
However, the errors were quite similar to the one's suggested in the comments
of protectFromC function in codegen.cpp. Thus, the fix for the errors obtained
is essentially along those lines.
* There were some cases in which the function name generated did not look nice,
in which cases the module name was not appended to the end of the function