Turn off direct on optimization for iterators
There is an assert that fails in lowerIterators if the direct on statement
applies. It may be that disabling the assert is sufficient, but I didn't want
to disrupt lowerIterators while making this change. In any case, the direct on
optimization can change an iterator that had one yield into one with more than
one yield - meaning that iterator inlining might no longer apply.