Enable computing histograms using reduce intents
A natural way to histogram something in parallel is something like this:

    const buckets = <a range or domain with an index per bucket>;
    var histo: [buckets] int;

    forall dataElement in <data to plot> with (+ reduce histo) {
      const bucket = <compute the bucket that dataElement falls into>;
      histo[bucket] += 1;
    }

    writeln(histo);

To make this work, we first equip chpl__sumType() in ChapelReduce.chpl
to handle arrays.

Then, a simple program like the first half of
   test/reductions/vass/histogram-1.chpl

would contain/execute more autoDestroys than autoCopies, corrupting memory.
The compiler changes in this PR fix that.

DETAILS

Turns out that when I generate an AST that MOVEs the result of an
array-returning function without autoCopy-ing it, then
returnRecordsByReferenceArguments() on this function does not fire
and things go off rail. Namely, the function itself and I think other
array-returning functions that it invokes get an extra autoDestroy
on the return value. This was the case for two 'svar's,
named shadowVar and shadowVarReduc, that stored the result of identity():

// before this change:

  redRef1->insertBefore("'move'(%S, identity(%S,%S))", // init
                        svar, gMethodToken, parentOp);

Adding autoCopy on the result resolves the issue.

As well as I replaced a PRIM_ASSIGN with a call to "=" on the result
of generate(). Since the reduction var already exists, we really want
assign into it, not bitcopy. To call "=", I had to put in "new CallExpr"
explicitly. Before I relied on insertAfter() to invoke new_Expr,
alas the latter does not handle calls to functions with operator names
like "=".