Add future showing race-y chameneos behavior with icc
My rewrite of chameneos-redux failed last night for intel, seemingly
because of removing the duplicate "metWithSelf" increment.  I.e.,
we seem to miss counting between 1 and 5 meetings typically when
compiling with icc -O2.  Replacing the double increment of 'metWithSelf'
seemingly causes the race to go away.

Equivalently, removing the double increment from the release version
shows the race as well.  Here, I've copied the release version, made
it a correctness test future, skipped it for everything except intel
and run it 10 times in a loop to statistically improve the chances
that at least one run will fail as a placeholder to look into this
issue.