Updated module printout to reflect inclusion of Reflection
This test is actually testing a broken feature in modules - that many
standard/package modules are being 'use'd by default due to the module
inheritance rules. Since we now use Reflection in Sort and Sort is used
in other standard modules included in ChapelStandard, we must include
Reflection. It's personal TODO of mine to fix this issue..