Ignore stride information for Block and Stencil's boundingBox.
This is a common way to create a distributed array:

```chpl
use BlockDist;

var Dom = {1..10, 1..10} by 2;
var Space = Dom dmapped Block(Dom);
var Data : [Space] int;
```

Prior to this commit this program would fail to compile. It would emit
an error message at a line in BlockDist complaining that it is illegal
to assign a strided domain to an unstrided domain (the boundingBox
initialization). This seems confusing at first, and ultimately puts the
burden on the user to create an unstrided version of their index space.

With PR #3778 it is now easy to cast a strided domain to an unstrided
domain. Instead of failing at compile-time, we can instead drop the
stride information and only take what we actually care about: the lo..hi
bounds.

Also updates Block's documentation to reflect that we ignore strides for
the boundingBox.