Improve optimize on clauses
It was bothering me that downEndCount created an active message that
wouldn't run in GASNet's AM handler. That means that in order to complete
a remote task, we create a new local task... which seems to be a lot of
rigamarole.

So, I investigated how to get downEndCount's AM to run as a fast execute
on.

While there, I realized that fast executeOn wrappers have start/finish
--cache-remote fences when in fact they cannot contain communication (or
else wouldn't be eligible for fast).  If they contain no communication,
there is no reason to invalidate/flush the cache. Also, along these
lines, there are some functions, such as allocation, that don't contain
communication but that are not suitable to run in a signal handler (to
run with a fast execute on). I updated optimizeOnClauses to distinguish
between these cases (no communication vs signal handler suitable) so that
the cache fences can be removed in more cases and so that the
optimization is clearer.