Fix a bug for recursive functions that return records
The "return record by reference argument" transformation applied the
transformation for a function, then later for all of the calls to that
function.  If the function is recursive, the recursive call would have
fewer actual arguments than the function had formals while transforming
the function body, leading to an actual/formal count mismatch error.

Transforming all of the calls before the function itself avoids this
problem.

Add a simple test that caused an internal error prior to this change.