Remove .bad files for futures, add .futures for failing initializers tests
Remove the .bad files because the expected failure output will be changing as I
go.  Add futures so that as these tests fail, it won't disrupt testing.  I'll
need to update which tests are futures and which aren't with each change, but
it'll keep the work in sync with master.