add a draft of a size function for enums
is a compiler-generated function for each enum that takes the enum and
returns the number of elements, without needing to iterate through the
tuple.

Is only a draft because I am very new to the compiler and may have
something that works for my test but breaks for some other things.