Allow cast from numeric types to imag
Before this change, this program:

  writeln(1:imag);

would output 0, under the thinking that a cast to :imag should pull out
the imaginary component, which is 0 for a real number.

However, this behavior has been surprising to several Chapel developers.
Furthermore, since it is an explicit cast, it seems that there is a low
likelihood of suprise from any inconsistency (namely - it doesn't produce
the same output as casting from real to complex and then to imag).

This PR allows explicit cast from real to imag to return a non-zero
value.