Remove the error for unresolved enum type symbols
This is my current way to permit .methods to be resolved for enums.
Unresolvable .methods will still be caught at the function resolution
pass.

Other options include allowing .size specifically, warning the user
when a symbol is allowed to pass through unresolved, or changing where
in the passes of the compiler enum symbol resolution occurs.