Cleaner generated code for local blocks
This commit attempts to improve the quality of the generated code for
local blocks. First, we will now emit a comment "/* local block */" to
make it clear when a local block begins. Second, we remove empty local
blocks at the beginning and towards the end of insertWideReferences.
These two changes should have the most positive impact on the generated
code.

The final transformation is to merge consecutive local blocks together.
Local blocks can be split up by the function 'fragmentLocalBlocks' in
lowerIterators. By the time we get to insertWideReferences, we find that
some of these local blocks are consecutive. One can also observe that
these local blocks are sometimes only separated by DefExprs. In that
case, we move those DefExprs before the earlier block when merging. This
transformation will not touch local blocks that were split across loops
or if-statements. For example, this:

```
{
  ...
  ('local block');
}
var len:int(64)
var call_tmp:int(64)
{
  ...
  ('local block');
}
```

Becomes this:
```
var len:int(64)
var call_tmp:int(64)
{
  ...
  ...
  ('local block')
}
```