Convert chameneosredux-race into a suppressed perf test
chameneosredux-race fails with the intel compiler, but only at the full problem
size and only on certain hardware. Brad added it as a correctness future that
was skipped for non-intel configurations. This meant it only ran on the
whiteboxes, where we didn't see the failure.

This converts it from a correctness future into a suppressed performance test.
The test will run once with the correctness problem size for correctness
testing. I expect that to pass everywhere.

It does performance testing with the full problem size which I expect it to
pass everywhere except the intel perf configuration. We do 25 trials to make
sure that the intel case fails, and to give us more confidence that it is an
intel specific issue.

25 trials may seem excessive, but the test only takes ~5 seconds on chap03 (our
slowest machine) so at most this is adding ~2 minutes of testing time.

See JIRA 226 for more info: https://chapel.atlassian.net/browse/CHAPEL-226