Move PRIM_WARNING removal to removeRandomPrimitive
I was seeing a PRIM_WARNING make it all the way to optimizeOnClauses for
the test test/arrays/bradc/arrOfDom/arrOfDom4.chpl.

I feel that the old logic in resolution in removeCompilerWarnings just
complicated the issue by using a table to decide which PRIM_WARNINGS to
remove when in fact all of them should be gone.

So, I removed removeCompilerWarnings and added a simple case to
removeRandomPrimitive that removes PRIM_WARNING and PRIM_ERROR calls.