Adjust cache to be re-entrant
When meaningful code is run as part of an executeOnFast, the cache calls
to chpl_comm, such as chpl_comm_wait_nb_some, can execute user code that
includes cache fences. In such cases, these fences should do nothing,
because a fast AM handler cannot communicate.

Here we guard the cache against this possibility.  A potentially better
alternative would be for the compiler to guarantee that no fences could
be called from an executeOnFast body.