Remove local blocks in Atomics, fix optimizeOnClauses for case
within optimizeOnClauses, the on body for atomic int . sub
is marked as not fast because of this:

    (CallExpr move
      (SymExpr 'tmp:_ref(atomic_int_least64_t)')
        (SymExpr 'call_tmp:__wide__ref_atomic_int_least64_t'))

However this operation is just creating a local ref from
a wide ref. This does not require communication.

Adjusts optimizeOnClauses to make the nearby logic clearer,
and to handle this case. Removes the local blocks from Atomics.chpl.