Recognize and permit super.init() and this.init() calls (but don't act on them)
Prior to this change, calls to super.init() and this.init() would resolve
weirdly.  With this change, they will not cause compiler errors, but will
basically be treated as no-ops, allowing many of my tests with no-op
super.init() calls in them to function properly.  My next step is to cause
these calls to function - allow super.init() when the type has a parent class
to call the parent's initializer, and allow this.init() calls to call another
initializer on the type.

IMPLEMENTATION DETAILS:

We wouldn't be able to trace the init() portion to the appropriate
_construct_type function.  The AST was transformed sufficiently that the "super"
or "this" portion was separated from the init call, so resolving the call in
function resolution on its own would've relied on a large amount of strange AST
knowledge that could easily break at any time.

With that in mind, I decided to add a couple of primitives to the compiler,
PRIM_SUPER_INIT and PRIM_THIS_INIT.  Doing this allowed me to keep the
"super"/"this" portion associated with the "init" portion of the call.
Recognizing when to insert these primitives was a bit unfortunate - I had to
do a strcmp against all the names involved to be sure we were entering the right
case.  I decided to do this instead of causing the compiler to flag these
names as keywords, which in the case of "this" and "super" would've meant
updating a lot of our rules and logic associated with those names to use the
keywords instead.  As that seemed unfortunate and costly, adding new primitives
felt more reasonable.

I decided to add these primitives during build.cpp instead of later in the
compiler (say, scope resolution or normalize).  This seemed the most reasonable
place, since later places in the compiler will involve more work to
backtranslate the changes that have happened to the AST, and would just end up
removing the nodes that were placed there to start.

Because these calls are created through the parser's dot_expr rule, however,
some AST shuffling was required - dot_exprs are inserted into the AST as
a call, stored in the baseExpr field on a call with the arguments (i.e.
``call ( call base . modifier) args`` ).  This meant that we needed to
transform the whole expression so that the arguments would be applied to the
super.init or this.init call, instead of having normalize move the call away
from the arguments because it and function resolution thought it needed to
return an object to work with.  To fix this, I had normalize recognize the
primitives and squash the call into a single level
( i.e. `` call super.init args `` )

Then, to allow this change to work without leaving junk in the compiler with
my stub implementation, I had functionResolution change the primitives into
PRIM_NOOP.  This was able to be done at precisely the point where I want to
have function resolution determine what to do about those calls, thankfully.