Add gatherWellKnownFns
filesToAST.cpp has historically included a list of Well Known Types that
must be found in parsing and that then are attached to global variables
in the compiler.

This patch adjusts filesToAST to also perform a similar operation
to gather Well Known Functions. Right now, there are 3 Well Known
Functions:

 - chpl_here_alloc
 - chpl_here_free
 - chpl_doDirectExecuteOn

but we expect others might be needed in the future.  One thing to note is
that this treatment for a function is only really necessary if calls to
that function are added after resolution. (Otherwise you could just use
new CallExpr(wellKnownFnName, ...) ). All of the functions here need
to be sometimes added after resolution (although chpl_doDirectExecuteOn
will be added after resolution only after the following commit).