Move direct-on-optimization to parallel.cpp
Mike correctly pointed out that the conditional added for the
direct-on-optimization would be more naturally added in parallel.cpp.
This change moves the code in question from createTaskFunctions.cpp
to a new function, createConditionalForDirectOn, in parallel.cpp.

In order to do so, chpl_doDirectExecuteOn needed to be resolved and
available to be added in new CallExprs after resolution. The previous
commit handles that part.