Never heap allocate argument bundles
The brave new world is that the runtime is responsible for copying
argument bundles. There does not seem to be a benefit to heap allocating
them when the runtime will just copy them from the stack or the heap
equally well. (Even GASNet active messages can be stack-allocated...).

One might argue that heap allocating the argument bundles saves valuable
stack space, but as much stack space would be used in the case that the
function is called directly (as with direct on optimization). Additionally,
we could structure the C code so that it is clear that this stack space
can be freed once the runtime call returns. (However it seems moot because
the arguments we are putting into the arg bundle are probably already
on the stack. E.g. if they are function arguments).

This resolves an issue I was seeing where a task function tried
to free its argument bundle, but the bundle the task fn gets is
a copy, not the original, so it core dumped. If we needed to heap
allocate some argument bundles for task functions, we would need
to include a pointer to the original arg bundle in the arg bundle
passed to the runtime.