Massivethreads updates for stack-arg-bundles
Massivethreads seems to include already a task-local storage mechanism in
the form of the 'custom_data' field. Use that because it copies the data,
where the void* argument does not.

With this change, CHPL_TASKS=massivethreads is able to run
the test/release/examples/hello*.chpl. The GASNet configuration
is not currently working but I think that's due to a different
issue. (myth_get_pthread_funcs: Assertion `real_pthread_create' failed.).
Since massivethreads+GASNet did not even compile before this change,
I'm not going to worry about that right now.