Update RA verification to be faster, use Timer
After PR #4008, this test was timing out in nightly testing in a slower
configuration. Some basic troubleshooting revealed that the verify step takes
much much longer than the kernel in the tested configuration.

This PR adjusts the verify step to use an on statement, which is much faster,
because it is usually working with a local lock.

While there, I updated the kernel to use a Timer instead of subtracting two
getCurrentTime numbers (noting that getCurrentTime returns the time since
midnight, so we would get a negative elapsed time if the test ran during
midnight).