Fix an issue where insertWideReferences attempted to analyze the wrong AST pattern.
The test memory/sungeun/refCount/arrays.chpl was failing intermittently
due to a compile-time assertion failure, and was challenging to reproduce.
Eventually @ronawho discovered that the intermittent nature was due to
the varying order of arguments to a function, something we believe to
be a side-effect of our homegrown map that hashes on pointers.

The problem was that IWS was looking for a pattern like this:

sym = foo(...);

Where foo returned a variable named 'cause'. Instead we were finding
the pattern:

bar(cause);
return cause;

Where bar looks like this:

fn bar(sym, ...) { ... }

To avoid this case, we test that the expr that caused 'sym' to be a
wide variable is used in a PRIM_RETURN.