Revert the timer changes made to RA in #4076
#4076 updated RA to use a Timer instead of getCurrentTime() since
getCurrentTime() resets at midnight. Unfortunately the RA comm count test uses
getCurrentTime() as a sentinel to figure out where to insert comm count code.

The right solution is probably to verify that the HPCC benchmarks aren't
required to use getCurrentTime(), update all of them to use a timer, and
finally update the  script that inserts the comm counting code. However, I
don't have a lot of time to look at that right now, so in order to quiet the
regression this just reverts the timer changes.