Only avoid using spool files for moab/torque systems
#4075 updated chpl_launchcmd to redirect stdout/stderr directly to the output
file on all pbs/aprun systems in order to avoid using spool files under
moab/toque systems.

This made executing on a PSBPro system really slow, because they normally
buffer things to /tmp on the compute node, and then at the end copy the output
directly to output file. This updates chpl_launchcmd to only redirect for
moab/torque systems so that we have the same behavior that we used to for
PBSPro systems.