Add submitted shootout entry for spectral-norm
Add the version of spectral-norm that we submitted to the submitted shootout
directory.

I basically copied the release version, double checked that it's the same
version we submitted, updated the .dat file specified in the .perfkeys file,
and added a .graph file that compares the release with the submitted version.