Checkout the master version of util/cron for the 1.13 shootout job
In addition to checking out the master test/ dir, we also need to checkout the
util/cron dir so that we can actually find the right script to run.

Note that a better solution would probably be to update nightly to either use
the 1.13 tarball directly, or to use a version of chapel installed on the
machine, but that's more work than I want to do right now.