Update tests for gmp
Add a test that runs (almost) all of the constructors for BigInt.

update the other tests to no longer delete the record BigInt, as it is
not a class anymore.

update how the dist_array test does initialization since the record
works a little differently than the class, and since the record’s
assignment requires both values to be refs (mpz_init_set requires
refs).