Fix memory leaks in BigInt assignment
auto copy, init copy and op= assignment no longer leak in most cases.

Added some simple tests — will refine them later, as the overlap
somewhat.

the gmp_dist_array test is still leaking, a lot. As far as I can tell,
none of the individual operations in it are leaking, it just is leaking
a bunch of things from the distributed array.