Add an OpenMP variant of the task spawning perf tests
Add an OpenMP comparison to the task spawning suite. This is the OpenMP
equivalent of our forall version. It's careful to only use the number of
physical cores so that we compare the same number of tasks being created. The
code to determine the number of physical cores was grabbed from
test/runtime/gbt/numThreadsSymbolicPhysical.prediff