Make reviewer suggested updates to complex math functions
Changed functions 'arg' and 'proj' to 'carg' and 'cproj' since they are only
for complex numbers and the names without the leading 'c' are more likely
to be used for other purposes.

Merged a few functions for different sized complexes into single functions.

Updated the wording of a chpldoc comment.