Remove warning for assigning serial iterator/iterable to array
Traditionally, we've generated warnings when assigning a serial
thing to an array.  Originally, I believe this was to mark cases
in our implementation where we hadn't yet implemented parallel
iterators for standard domain maps.  However, over time, it's
become more of an annoyance as things like the following:

       var A: [1..10] int = for i in 1..10 do i;

shouldn't generate a warning -- the user clearly asked for a
serial initializer for the array and shouldn't be warned
when doing so.  This came up in some recent shootout work.

Related to other similar serialization warning removals I've been
doing, I think the right way to get warnings for the original case now
is to implement a serial leader/follower iterator and have it generate
the warning rather than making the warning occur for any iterator that
is serial.  I.e., it ought to be legal and warning-free to serially
assign to, or initialize, an array.

To that end, removed the warning and updated .good files that
generated it, verifying that it didn't seem bad for these test
cases not to warn in the cases where they had been.