Add -nl 1 to PREEXEC files for string shootouts
For the string shootouts, we run fasta in order to generate input
decks.  Though these are designed to be single-locale tests, I was
accidentally testing them with gasnet enabled, which was breaking
because of the lack of an -nl argument.  Here I add -nl 1 to the
execution lines to avoid this issue in the future (while doing no harm
in the single-locale context).