Add --n=0 to all string *EXECOPTS to mimic shootout Makefiles
The official shootout Makefiles pass an unnecessary '0' (or, for
Chapel, --n=0) argument to the string benchmarks which we need to
consume and ignore.  This could be done either by adding a fake config
or by having an args argument on main().  We chose the latter because
it seems more innocuous (and is more similar to how languages like C
arguably drop such arguments on the floor).  Here, I'm adding --n=0 to
all of our EXECOPTS files in order to make sure we don't accidentally
remove this support again in the future.