Avoid a warning from clang about calling 'creal' without capturing the result
A bit of code in IO.chpl calls cplx.re.type, which ends up generating a call
to creal() without capturing the result since the use of the result was only
for a .type expression and was removed after function resolution.

Instead of fixing the underlying problem, just avoid it in this one spot. A
better fix would be to not evaluate the rest of the expression for .type calls.