Added/updated tests for GMP functions
Added tests that touch each function in the GMP library, in the process
discovered that some of the functions don’t work. These tests leak some
glom string data, but not any gmp data.

Updated gmp_dist_array to manually free the records since it leaks the
array elements (so it doesn’t leak 6 *10^6 bytes of gmp data).

Removed the freeing of BigInts in the BigInt versions of pidigits now
that BigInt is a record.