Link users guide to master branch
As I've been working on users guide edits, I realized that it's a
little regrettable that we didn't have the "work in progress" note
link to the master branch of the docs so that people could see the
latest version of the users guide easily.  This is an attempt to
remedy the situation for the next release (or maybe to back-patch
into the 1.13 docs).

I also did some wordsmithing to tighten up the note while here.