Turn series of if statements into a loop in meteor's fillContiguousSpace
Paul pointed out that the reason I was running into problems with that commented
out for loop was due to a typo.  I had thought I created the for loop using a
direct copy of the if statements involved to avoid this issue, but apparently
not.

Turned the anonymous array of values into an anonymous tuple to avoid a minor
performance hit.