Stop skipping gmp-chudnovsky for non-GNU compilers
As best I can tell, this skipif predates general support for GMP
and we should really be skipping based on gmp support being
available or not (which is already handled by the gmp.skipif a
few directories up).  So retiring this.

The presence of this .skipif is why we saw failures for gnu-over-clang
on Mac testing last night when we've never seen them for clang testing.
The test simply wasn't being run for clang testing.