Always squash C warnings, not just when we ask for "more" warnings
David and I recently realized that, surprisingly, certain tests
(exercises/Mandelbrot/mandelbrot.chpl was a good example when compiled
with clang before his PR #4132 yesterday) would generate C-based
warnings when compiled by default (chpl foo.cpl) but not when compiled
with C warnings turned on (chpl --cc-warnings foo.chpl). This seems
completely counterintuitive.  It turns out that the reason is that we
only throw certain -Wno* flags when --cc-warnings is requested, the
rationale (presumably) being that we turn on a bunch of warnings with
-Wall, but then there are some we need to turn back off based on the
code we generate today.  It may be that some of these warning flags
have gotten turned on by default in newer versions of compilers, or
that we've become more sensitive to them over time.

In any case, it doesn't seem good to get more warnings with
--no-cc-warnings than with --cc-warnings...  Nor to have users running
into warnings that our testing (which always throws --cc-warnings)
doesn't reveal.

To that end, this commit refactors our Makefiles so that compiler
flags that are used to squash warnings in the generated code are
separated out into their own Makefile variable
(SQUASH_WARN_GEN_CFLAGS) which is always added to the 'cc' command
line regardless of whether or not --cc-warnings is used (and after
WARN_GEN_CFLAGS so that it continues to override any warnings that are
turned on there).