Use jemalloc with PGI
Try using jemalloc with pgi (round 3? maybe round 4?)

#3944 was the latest attempt to use jemalloc with pgi, but we ran into some
segfaults and timeouts as a result a result of a jemalloc bug when TLS wasn't
available (such as under pgi.) That bug was fixed with 4.2.1 and the previous
red-black tree segfaults we saw have been fixed with 5e22f84. With those two
fixes, I think we can finally use jemalloc with pgi.