Clean up comments
removed the .localize() comments, since assignment localizes and multi
locale correctness testing succeeded.

Still have a lot of TODOs:
* maybeCopy is no longer used at all, and localize is not called at
all, but a user might want a way to localize without assignment?
Localize as currently written doesn’t allow the user to get a usable
shallow copy since it generally is used as tempBigInt =
remoteBigInt.localize(), so even if they are on the same locale the
assignment makes a deep copy of whatever localize returns.