Added operator overloads, add mpz_get_limbn, formatting
Added operator overloads for **, *, /, %, unary +, unary -, &, ^, |, +,
-, >=. <=, >, >, ==, and !=. All can be used with BigInt operands, some
of the two-operand ones can also be used with the second operand as a
uint or int.

Also removed a bunch of extra spaces from the ends of lines.

Removed the comment about hoping to eventually have this class be a
record, and to eventually support operator overloads.

Also added the extern for mpz_get_limbn, and updated the misc test and
good file to test it.