Remove the COMPOPTS file left over from an old strategy
The compopts file was to verify that these tests work under both branches of
insertDynamicDispatchCalls.  It was necessary to throw
--conditional-dynamic-dispatch-limit with a high value to enter the other
branch and verify that both worked.  However, it seems as if mere detection of
the super call within the "dynamic dispatch" branch (as opposed to the "virtual
method dispatch" branch) allows these tests to pass.  So, my new solution is
to always utilize that path if a super call is involved - this will be
implemented properly in the next commit.